Solve exercise 07-flex-layout-2
Add a .section div around the .sidebar and .card divs
and then a .container div around the .card divs.

This is to align the body in a column and then the .section in a row and
the .container also in a row but with flex-wrap.

Then add the necessary styles to recreate the exercise page.